Fix:Types field, move users_model to modles folder
