+++ app/api/serializers/uni_stu_feedback.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from ...additional_models.uni_stu_feedback import UniStuFeedback
+from ...models.uni_stu_feedback import UniStuFeedback
 
 
 class UniStuFeedback(serializers.ModelSerializer):
